changed logits to float64 to increase numerical stability

+++ losses/loss.py
@@ -26,10 +26,10 @@ class ListWiseLoss(Loss):
     def loss_per_list(self, y_true, y_pred, sample_weights=None):
         with tf.GradientTape() as g:
             g.watch(y_pred)
-            ignore_mask = tf.cast(y_true == -100, 'float32') #-100 is the default ignore value
+            ignore_mask = tf.cast(y_true == -100, y_pred.dtype) #-100 is the default ignore value
             use_mask = 1.0 - ignore_mask
-            noise =  ignore_mask * tf.random.uniform(y_pred.shape, 0.0, 1.0) 
-            listwise_ytrue = use_mask * tf.cast(y_true, 'float32') + noise
+            noise =  ignore_mask * tf.random.uniform(y_pred.shape, 0.0, 1.0, dtype=y_pred.dtype) 
+            listwise_ytrue = use_mask * tf.cast(y_true, y_pred.dtype) + noise
             listwise_loss = self.calc_per_list(listwise_ytrue, y_pred)
             use_loss_mask = tf.squeeze(use_mask[:,:1], axis=1)
             average_loss =  tf.reduce_sum(listwise_loss * use_loss_mask) / tf.reduce_sum(use_loss_mask)
